Cache instrumented object files

+++ include/Instrumentation/DynamicCallTree.h
@@ -16,7 +16,7 @@ namespace mull {
   struct CallTree {
     llvm::Function *function;
     int level;
-    uint64_t functionsIndex;
+    uint32_t functionsIndex;
     std::list<std::unique_ptr<CallTree>> children;
     CallTree(llvm::Function *f) : function(f), level(0), functionsIndex(0) {}
   };
@@ -34,19 +34,19 @@ public:
   DynamicCallTree() = delete;
   ~DynamicCallTree() = delete;
 
-  static std::unique_ptr<CallTree> createCallTree(uint64_t *mapping, std::vector<CallTreeFunction> functions);
+  static std::unique_ptr<CallTree> createCallTree(uint32_t *mapping, std::vector<CallTreeFunction> functions);
   static std::vector<CallTree *> extractTestSubtrees(CallTree *root, Test *test);
   static std::vector<std::unique_ptr<Testee>> createTestees(std::vector<CallTree *> subtrees,
                                                             Test *test,
                                                             int distance,
                                                             Filter &filter);
 
-  static void enterFunction(const uint64_t functionIndex,
-                            uint64_t *mapping,
-                            std::stack<uint64_t> &stack);
-  static void leaveFunction(const uint64_t functionIndex,
-                            uint64_t *mapping,
-                            std::stack<uint64_t> &stack);
+  static void enterFunction(const uint32_t functionIndex,
+                            uint32_t *mapping,
+                            std::stack<uint32_t> &stack);
+  static void leaveFunction(const uint32_t functionIndex,
+                            uint32_t *mapping,
+                            std::stack<uint32_t> &stack);
 };
 
 }
